projet_obs/requirements.txt
requests
obsws-python
python-dotenv